Extract node lock-handling code into separate file, start writing tests
